// Frontend/AngularHealthApp/src/app/models/patient.model.ts
export class Patient {
  id?: string;
  title?: string;
  firstname?: string;
  surname?: string;
  dob?: Date;
  gender?: string;
  address1?: string;
  address2?: string;
  town?: string;
  county?: string;
  postcode?: string;
  telephoneNumber?: string;
  mobileNumber?: string;
  email?: string;
  deleted?: boolean;
}